+++ bank-feeds/src/codatbankfeeds/sdkconfiguration.py
@@ -2,9 +2,10 @@
 
 import requests
 from dataclasses import dataclass
-from typing import Dict, Tuple
+from typing import Dict, Tuple, Callable, Union
 from .utils.retries import RetryConfig
 from .utils import utils
+from codatbankfeeds.models import shared
 
 
 SERVERS = [
@@ -16,14 +17,14 @@ SERVERS = [
 @dataclass
 class SDKConfiguration:
     client: requests.Session
-    security_client: requests.Session
+    security: Union[shared.Security,Callable[[], shared.Security]] = None
     server_url: str = ''
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '4.0.0'
-    gen_version: str = '2.159.2'
-    user_agent: str = 'speakeasy-sdk/python 4.0.0 2.159.2 3.0.0 codat-bankfeeds'
+    sdk_version: str = '5.0.0'
+    gen_version: str = '2.195.2'
+    user_agent: str = 'speakeasy-sdk/python 5.0.0 2.195.2 3.0.0 codat-bankfeeds'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: